Work in progress improvements
