+++ src/api/Networking/Player.cs
@@ -1,4 +1,5 @@
-﻿using CSM.API.Networking.Status;
+﻿using System;
+using CSM.API.Networking.Status;
 using LiteNetLib;
 
 namespace CSM.API.Networking
