+++ src/basegame/CSM.BaseGame.csproj
@@ -220,6 +220,19 @@
     <Compile Include="Injections\PropHandler.cs" />
     <Compile Include="Injections\RoadHandler.cs" />
     <Compile Include="Injections\TerrainHandler.cs" />
+    <Compile Include="Injections\Tools\BaseToolHandler.cs" />
+    <Compile Include="Injections\Tools\BuildingToolHandler.cs" />
+    <Compile Include="Injections\Tools\DefaultToolHandler.cs" />
+    <Compile Include="Injections\Tools\NetToolHandler.cs" />
+    <Compile Include="Injections\Tools\PlayerCursorManager.cs" />
+    <Compile Include="Injections\Tools\PropToolHandler.cs" />
+    <Compile Include="Injections\Tools\TerrainToolHandler.cs" />
+    <Compile Include="Injections\Tools\ToolCommandBase.cs" />
+    <Compile Include="Injections\Tools\ToolManagerOverlayRenderer.cs" />
+    <Compile Include="Injections\Tools\ToolSimulatorCursorManager.cs" />
+    <Compile Include="Injections\Tools\TransportToolHandler.cs" />
+    <Compile Include="Injections\Tools\TreeToolHandler.cs" />
+    <Compile Include="Injections\Tools\ZoneToolHandler.cs" />
     <Compile Include="Injections\TransportHandler.cs" />
     <Compile Include="Injections\TreeHandler.cs" />
     <Compile Include="Injections\WeatherHandler.cs" />
