+++ src/csm/CSM.cs
@@ -2,9 +2,11 @@
 using CitiesHarmony.API;
 using ColossalFramework.IO;
 using CSM.API;
+using CSM.API.Networking;
 using CSM.Commands;
 using CSM.Injections;
 using CSM.Mods;
+using CSM.Networking;
 using CSM.Panels;
 using ICities;
