+++ src/csm/CSM.csproj
@@ -114,19 +114,6 @@
     <Compile Include="Injections\PauseMenuHandler.cs" />
     <Compile Include="Injections\SpeedPauseHandler.cs" />
     <Compile Include="Injections\TickLoopHandler.cs" />
-    <Compile Include="Injections\Tools\ToolCommandBase.cs" />
-    <Compile Include="Injections\Tools\ToolManagerOverlayRenderer.cs" />
-    <Compile Include="Injections\Tools\BuildingToolHandler.cs" />
-    <Compile Include="Injections\Tools\ToolSimulatorCursorManager.cs" />
-    <Compile Include="Injections\Tools\PlayerCursorManager.cs" />
-    <Compile Include="Injections\Tools\DefaultToolHandler.cs" />
-    <Compile Include="Injections\Tools\TreeToolHandler.cs" />
-    <Compile Include="Injections\Tools\BaseToolHandler.cs" />
-    <Compile Include="Injections\Tools\NetToolHandler.cs" />
-    <Compile Include="Injections\Tools\TerrainToolHandler.cs" />
-    <Compile Include="Injections\Tools\PropToolHandler.cs" />
-    <Compile Include="Injections\Tools\TransportToolHandler.cs" />
-    <Compile Include="Injections\Tools\ZoneToolHandler.cs" />
     <Compile Include="Models\ColorSurrogate.cs" />
     <Compile Include="Models\QuaternionSurrogate.cs" />
     <Compile Include="Models\ControlPointSurrogate.cs" />
