+++ src/csm/Commands/Handler/Internal/ClientDisonnectHandler.cs
@@ -4,8 +4,8 @@ using CSM.API.Networking;
 using CSM.BaseGame.Helpers;
 using CSM.Commands.Data.Internal;
 using CSM.Networking;
-using CSM.Injections.Tools;
 using ColossalFramework;
+using CSM.BaseGame.Injections.Tools;
 
 namespace CSM.Commands.Handler.Internal
 {
